requirements.txt
google-generativeai
langchain
langchain-google-genai
streamlit
chromadb
pandas